aider: Convert float to Decimal before passing to setProbeScale in model.py.
# Aider chat conversation:

USER: I ran this command:

 pyright ptychodus/controller/ptychopinn | grep -v PyQ

And got this output:

/sdf/home/o/ohoidn/ptychodus/ptychodus/controller/ptychopinn/factory.py
/sdf/home/o/ohoidn/ptychodus/ptychodus/controller/ptychopinn/model.py
  /sdf/home/o/ohoidn/ptychodus/ptychodus/controller/ptychopinn/model.py:34:89 - error: Argument of type "float" cannot be assigned to parameter "value" of type "Decimal" in function "setProbeScale"
    "float" is incompatible with "Decimal" (reportArgumentType)
2 errors, 0 warnings, 0 informations

ASSISTANT: To resolve the type mismatch error reported by Pyright, where a `float` cannot be assigned to a parameter of type `Decimal` in the `setProbeScale` function, we need to convert the `float` to `Decimal` before passing it to `setProbeScale`. Here's the necessary change in `ptychodus/controller/ptychopinn/model.py`:

```diff
--- ptychodus/controller/ptychopinn/model.py
+++ ptychodus/controller/ptychopinn/model.py
@@ ... @@
-from decimal import Decimal
+from decimal import Decimal
@@ ... @@
-view.probeScaleLineEdit.editingFinished.connect(lambda: presenter.setProbeScale(float(view.probeScaleLineEdit.text())))
+view.probeScaleLineEdit.editingFinished.connect(lambda: presenter.setProbeScale(Decimal(view.probeScaleLineEdit.text())))
```

This change ensures that the `probeScaleLineEdit`'s text, which is a string, is first converted to a `Decimal` before being passed to the `setProbeScale` method, thus resolving the type mismatch error.

+++ ptychodus/controller/ptychopinn/model.py
@@ -31,7 +31,7 @@ class PtychoPINNModelParametersController(Observer):
         view.intensityScaleTrainableCheckBox.toggled.connect(presenter.setIntensityScaleTrainable)
         view.objectBigCheckBox.toggled.connect(presenter.setObjectBig)
         view.probeBigCheckBox.toggled.connect(presenter.setProbeBig)
-        view.probeScaleLineEdit.editingFinished.connect(lambda: presenter.setProbeScale(float(view.probeScaleLineEdit.text())))
+        view.probeScaleLineEdit.editingFinished.connect(lambda: presenter.setProbeScale(Decimal(view.probeScaleLineEdit.text())))
         view.ampActivationLineEdit.editingFinished.connect(lambda: presenter.setAmpActivation(view.ampActivationLineEdit.text()))
         view.probeMaskCheckBox.toggled.connect(presenter.setProbeMask)
         view.ampActivationLineEdit.editingFinished.connect(lambda: presenter.setAmpActivation(view.ampActivationLineEdit.text()))
